% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageTCGAWorkflow-package.R
\docType{package}
\name{imageTCGAWorkflow-package}
\alias{imageTCGAWorkflow}
\alias{imageTCGAWorkflow-package}
\title{imageTCGAWorkflow: Workflow for Histopathology Image Analysis with TCGA Data}
\description{
A Bioconductor workflow demonstrating end-to-end histopathology image
analysis using TCGA diagnostic whole-slide images. See the package vignette
for the complete workflow.
}
\seealso{
\code{vignette("imageTCGAWorkflow", package = "imageTCGAWorkflow")}
}
\author{
\strong{Maintainer}: Ilaria Billato \email{ilaria.billato@unipd.it} (\href{https://orcid.org/0000-0002-3335-3254}{ORCID}) (Department of Biology, University of Padova)

}
\keyword{internal}
